[ticket/15538] Split macros and add docblocks
PHPBB3-15538
